requirements.txt
# Core dependencies - Python 3 compatible versions
langchain==0.0.312
langchain-openai==0.0.3
openai==1.3.5
python-dotenv==1.0.0

# Document processing
pypdf==3.16.2
python-docx==1.0.1
unstructured==0.10.30
nltk==3.8.1

# Vector database
chromadb==0.4.18
sentence-transformers==2.2.2

# API and web
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
flask-cors==3.0.10

# Frontend
streamlit==1.28.1
pillow==10.1.0
pandas==2.1.3
requests==2.31.0

# Testing
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
